[PRDP-104] updated receipt model with notification data
