--- src/main/java/it/gov/pagopa/receipt/pdf/service/model/EventData.java
@@ -1,13 +1,0 @@
-package it.gov.pagopa.receipt.pdf.service.model;
-
-import lombok.*;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class EventData {
-    private String payerFiscalCode;
-    private String debtorFiscalCode;
-    private String transactionCreationDate;
-}
