--- src/main/java/it/gov/pagopa/receipt/pdf/service/model/ReasonError.java
@@ -1,13 +1,0 @@
-package it.gov.pagopa.receipt.pdf.service.model;
-
-import lombok.*;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReasonError {
-    private int code;
-    private String message;
-
-}
