+++ src/main/java/it/gov/pagopa/receipt/pdf/service/model/receipt/EventData.java
@@ -1,0 +1,21 @@
+package it.gov.pagopa.receipt.pdf.service.model.receipt;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class EventData {
+    private String payerFiscalCode;
+    private String debtorFiscalCode;
+    private String transactionCreationDate;
+    private String amount;
+    private List<CartItem> cart;
+
+}
