+++ src/main/java/it/gov/pagopa/receipt/pdf/service/service/impl/AttachmentsServiceImpl.java
@@ -6,8 +6,8 @@ import it.gov.pagopa.receipt.pdf.service.enumeration.AppErrorCodeEnum;
 import it.gov.pagopa.receipt.pdf.service.exception.*;
 import it.gov.pagopa.receipt.pdf.service.model.Attachment;
 import it.gov.pagopa.receipt.pdf.service.model.AttachmentsDetailsResponse;
-import it.gov.pagopa.receipt.pdf.service.model.Receipt;
-import it.gov.pagopa.receipt.pdf.service.model.ReceiptMetadata;
+import it.gov.pagopa.receipt.pdf.service.model.receipt.Receipt;
+import it.gov.pagopa.receipt.pdf.service.model.receipt.ReceiptMetadata;
 import it.gov.pagopa.receipt.pdf.service.service.AttachmentsService;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
