+++ src/test/java/it/gov/pagopa/receipt/pdf/service/service/impl/AttachmentsServiceImplTest.java
@@ -7,7 +7,11 @@ import it.gov.pagopa.receipt.pdf.service.client.ReceiptCosmosClient;
 import it.gov.pagopa.receipt.pdf.service.enumeration.AppErrorCodeEnum;
 import it.gov.pagopa.receipt.pdf.service.exception.FiscalCodeNotAuthorizedException;
 import it.gov.pagopa.receipt.pdf.service.exception.InvalidReceiptException;
-import it.gov.pagopa.receipt.pdf.service.model.*;
+import it.gov.pagopa.receipt.pdf.service.model.Attachment;
+import it.gov.pagopa.receipt.pdf.service.model.AttachmentsDetailsResponse;
+import it.gov.pagopa.receipt.pdf.service.model.receipt.EventData;
+import it.gov.pagopa.receipt.pdf.service.model.receipt.Receipt;
+import it.gov.pagopa.receipt.pdf.service.model.receipt.ReceiptMetadata;
 import it.gov.pagopa.receipt.pdf.service.service.AttachmentsService;
 import jakarta.inject.Inject;
 import lombok.SneakyThrows;
